Productionize the threaded libjuju and update library
This patch updates the sync_wrapper() function to fully be able to run
async functions that interact with libjuju in a background thread to be
able to leave the async coroutines running that libjuju launches when
connecting to a controllers.

The rest of the patch is updating the library, and tests, to work with
the updated sync_wrapper().